requirements.txt
--extra-index-url=http://pypi.fum.futurice.com/simple/
Django==1.5.5
psycopg2==2.5.2
python-dateutil==2.1
django-extensions==1.2.0
python-ldap==2.4.13
MySQL-python==1.2.4
pytz==2013d
Fabric==1.6.1
watchdog==0.6.0
assetgen==0.3.4
Werkzeug==0.9.1
Markdown==2.3.1
django-filter==0.6
djangorestframework==2.3.9
django-rest-framework-docs==0.1.6
django-crispy-forms==1.3.2
django-haystack==2.0.0
django-ignoretests==0.3.2
pysolr==3.0.6
lxml==3.2.3
smbpasswd==1.0.2
mock==1.0.1
PIL==1.1.7
python-memcached==1.53
diff-match-patch==20120106
ldappool==1.0
South==0.8.2
django_js_utils_ng==0.4.futu

django-history==0.1.6
mockldap==0.1.4futu
slumber==0.6.5.futu
